changed enum STAFF_TYPE to strong enum StaffType, moved it to Peep.h

+++ src/openrct2/peep/Peep.cpp
@@ -1940,7 +1940,7 @@ void Peep::FormatNameTo(Formatter& ft) const
                 STR_ENTERTAINER_X,
             };
 
-            auto staffNameIndex = AssignedStaffType;
+            auto staffNameIndex = static_cast<uint8_t>(AssignedStaffType);
             if (staffNameIndex > sizeof(staffNames))
             {
                 staffNameIndex = 0;
